Fix bug with initial accidental
The discrimination between Roman numeral and note name and the detection
of accidentals on either is tricky.  Inadvertently, a bug was introduced
in the process of simplifying Scheme functions.  This commit fixes the
problem.

Detection of initial accidental now works with figured-bass numbers.

Figured-bass numbers are now arranged as a right-justified column so altered
figures will be aligned properly.  (Case of an alteration _following_ a
number is not considered.  Will this ever be supported?)

Procedure inversion? introduced to check for figured-bass numbers.

Added some docstrings.

Small changes to README.