Improvements in note-name handling
To identify a note-name, we simply compare a string with a list derived
from the list of all note-names found in scm/define-note-names.scm.  Though
capital letters are not allowed in LilyPond input, we accept them here to
allow lowercase and uppercase names in the output as indications of
chord quality.

Input is now possible with any of the English note-names, and any English
alteration.

Comments and README are updated to reflect these changes.