Avoid superfluous stencil expressions
As much as possible, restructure code so empty markups and translations
of 0.0 don't find their way into the stencil expression.

It does not appear that any extra kerning is needed after alterations.
(Note commenting-out in big-char?)  Because of this, a make-hspace-markup
call was removed from make-base-markup.

Add an appropriate self-alignment-X to example.

Rename divide-by-spaces to divide-at-spaces.