Move font-size logic for figures, renaming, comments
Font-size logic has been moved from make-figures-markup, which is
concerned with building columns of figures, to format-figures,
which creates the elements of the column.

Inter-figure spacing is now controlled by the font-size of the
figures rather than the global font-size.

A number or variables and functions have been renamed for clarity:
second-part -> secondary-markup, big-char? --> get-extra-kerning,
inversion --> figures, size --> font-size

Comments added to clarify method of vertical alignment of
quality marker and figures.

Rearrangement of operands in multiplications involving scaling-factor.