Further integration of scaleDegree
Use procedures already in place for parsing string and
returning accidental markup.

This has the benefit of allowing any of the English accidentals
as a prefix, not that sharps, flats, and naturals aren't enough...

TODO: How do we deal sensibly with non-accidental modifiers?  Right
now if an accidental prefix is not found, it is grouped with the
number.  Thus, the caret is centered over both number and
unrecognized character(s).