Flat is not a quality
Previously notations like this were possible:

\markup \rN { V f 9 }

This doesn't make sense.  The flat obviously belongs with the "9:
either "f9" is a marker of quality, or it designates an altered
figure.  Right now figures are distinguished solely by the
inclusion of numbers, so "f9" is automatically designated a
figure.

As a result of this patch, "f" above will not be rendered as
a flat.  \markup \rN { V f9 } is needed to represent a Mmm9 chord.

(This makes the input more consistent: \rN { V s 7 } was not possible,
for example.)

One issue that has not been resolved is the precise difference
between a quality marker and a figure, possibly altered.

A problem case is the half-diminished marker, which must be used in
conjunction with "7". The question is: is the struck symbol the
indicator of quality, which must appear with a _figure_ "7" (after all,
they are clearly separate when the seventh chord is _inverted_); or
is the quality marker either a compound one -- "h7" or separable?

(This seems pedantic, but it has consequences for how the parsing
algorithms are designed.)